+++ tests/valid-gd-scripts/annotation-w-empty-body-54279.gd
@@ -1,1 +1,2 @@
+extends Node
 @onready() var x